requirements.txt
# TTS==0.20.6
flask
flask-cors
llama_cpp_python
PySide6
keyboard